BIG commit!
would not normally push this big of a commit but am in a time crunch. Added 4 new handlers those being toggletask, edittask, viewtask, and showtaskdetails. These are used for different htmx interactions in the application. Toggle task is for the htmx of the checkbox. Edit task is used for the htmx logic when clicking the edit button and editing a task. View task is the htmx logic for the collapsed view of the task. Show tasks details handles the htmx logic when clicking to expand the task and view description. I also had to change the output of each CRUD operation since before they returned JSON but I found out HTMX needs to return HTML template. Added styles.css this was copied from remarqable website (sorry if this was not allowed). For templates added index.html, task-details.html, task-edit.html, task-item.html, and tasks.html. index.html is the main page html. task-details is the template for expanding the task description. task-edit is the template for rendering the edit form. task-item is the template for the compact task view. tasks.html is the template for rendering all task from the get tasks handler. I still need to implement the strike through which will be another handler, route, template, and css adjustments. This will include adding a seperate area below for completed tasks.